requirements.txt
pyvespa>=0.45.0
numpy>=1.24.0
